Android Port Tests now run in a single qemu
instead of calling "run" for each android port test, we pass multiple --shell-command
options to a single call of "run". this starts just a single qemu session and saves
lots of time.

I also added in the capability for standard Android Tests to be groupled in a similar
way to the "androidporttests" are grouped with a new keyword in the test-map:
"androidtests".

Bug: 234727361
Change-Id: Idc97837c06ea7967a3039a6acd14eb672f7e6f2d