Android Port Tests now run in a single qemu
instead of calling "run" for each android port test, we pass multiple --shell-command
options to a single call of "run". this starts just a single qemu session and saves
lots of time.

I also added in the capability for standard Android Tests to be groupled in a similar
way to the "androidporttests" are grouped with a new keyword in the test-map:
"androidtests". this isn't completely working right now, but I was encouraged to
commit and start getting feedback.

Change-Id: I4ae986f0d265658f06546575bd8d23137f6094d3