scripts:Run androidporttests in a single QEMU instance
It takes quite a while to start up a QEMU instance, we want
to run androidporttests within a single statement in the test-map
within a single call to QEMU.py rather than calling and setting
up android for each porttest.

Because, grouping the androidporttests into a single TrustyTest
object, we diable this feature when selecting a single test
to run with the build.py --test option.

Bug: 234727361
Change-Id: I5538369d555c745a035520fceb6284a6ada06674