scripts:apend "run" to TrustyTest obj in run_tests.py
previously "run --headless --shell-command" or
"run --headless --boottest" were added to the TrustyTest objects
during the reading of the build-config/test-map files. this
makes it difficult to decide later if the tests should be sent
down to the qemu together with one call to the "run" script or
independently. The run_tests.py module really controls and
knows how the "run" script works and the formation of the
run commandline should be done there.

Bug: 234727361
Change-Id: I1ae62edf8cad8abbbf84426966c3a8fc63db2e13